move HealthChecks.Azure.Data.Tables out of CosmosDb package
