+++ .github/workflows/healthchecks_cosmosdb_cd.yml
@@ -23,7 +23,11 @@ jobs:
         run: dotnet restore ./src/HealthChecks.CosmosDb/HealthChecks.CosmosDb.csproj
       - name: Build
         run: dotnet build --no-restore ./src/HealthChecks.CosmosDb/HealthChecks.CosmosDb.csproj -c $BUILD_CONFIG
+      - name: Pack Tables
+        run: dotnet pack --no-build ./src/HealthChecks.Azure.Data.Tables/HealthChecks.Azure.Data.Tables.csproj -c $BUILD_CONFIG -o ./artifacts
       - name: Pack
         run: dotnet pack --no-build ./src/HealthChecks.CosmosDb/HealthChecks.CosmosDb.csproj -c $BUILD_CONFIG -o ./artifacts
+      - name: Publish Tables
+        run: dotnet nuget push ./artifacts/AspNetCore.HealthChecks.Azure.Data.Tables.*.nupkg -k ${{secrets.NUGET_API_KEY}} -s https://api.nuget.org/v3/index.json --skip-duplicate
       - name: Publish
         run: dotnet nuget push ./artifacts/AspNetCore.HealthChecks.CosmosDb.*.nupkg -k ${{secrets.NUGET_API_KEY}} -s https://api.nuget.org/v3/index.json --skip-duplicate
