+++ src/HealthChecks.Azure.Data.Tables/HealthChecks.Azure.Data.Tables.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <PackageTags>$(PackageTags);Azure;Tables</PackageTags>
+    <Description>HealthChecks.CosmosDb is the health check package for Azure Tables.</Description>
+    <VersionPrefix>$(HealthCheckCosmosDb)</VersionPrefix>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Include="../ClientCache.cs" Link="ClientCache.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="7.0.9" />
+    <PackageReference Include="Azure.Data.Tables" Version="12.8.1" />
+  </ItemGroup>
+
+</Project>
