+++ src/HealthChecks.Azure.Data.Tables/TableServiceHealthCheckOptions.cs
