+++ src/HealthChecks.CosmosDb/HealthChecks.CosmosDb.csproj
@@ -8,9 +8,15 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Include="../ClientCache.cs" Link="ClientCache.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Microsoft.Azure.Cosmos" Version="3.35.3" />
-    <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="7.0.9" />
-    <PackageReference Include="Azure.Data.Tables" Version="12.8.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\HealthChecks.Azure.Data.Tables\HealthChecks.Azure.Data.Tables.csproj" />
   </ItemGroup>
 
 </Project>
