+++ src/HealthChecks.CosmosDb/Properties/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+using System.Runtime.CompilerServices;
+using HealthChecks.CosmosDb;
+
+[assembly: TypeForwardedTo(typeof(TableServiceHealthCheck))]
+[assembly: TypeForwardedTo(typeof(TableServiceHealthCheckOptions))]
